+++ CHANGELOG.md
@@ -31,6 +31,8 @@
 - `txOutRefToTxSkelOut` helper to query the mock chain for recreating a
   `TxSkelOut` from a `TxOutRef`. This is very useful when using Tweaks that need
   to pay back an existing output with a slight modification.
+- A new tweak `modifySpendRedeemersOfTypeTweak` to apply an optional
+  modification of all redeemers of a certain type within the skeleton inputs.
 
 ### Removed
 
@@ -61,6 +63,8 @@
   * it is not visible from outside of `cooked-validators`
 - Dependency to cardano-api bumped to 8.46.
 - The whole testing API has been revamped
+- File `AddInputsAndOutputs.hs` has been split into `Inputs.hs`, `Outputs.hs`
+  and `Mint.hs`.
 
 ### Fixed